+++ vendor-service/src/main/java/com/fastline/vendorservice/application/VendorService.java
@@ -1,17 +1,24 @@
 package com.fastline.vendorservice.application;
 
-import com.fastline.vendorservice.domain.entity.Order;
-import com.fastline.vendorservice.domain.entity.Product;
-import com.fastline.vendorservice.domain.entity.Vendor;
+import com.fastline.common.exception.CustomException;
+import com.fastline.common.exception.ErrorCode;
+import com.fastline.vendorservice.application.service.HubClient;
+import com.fastline.vendorservice.application.service.UserClient;
+import com.fastline.vendorservice.domain.model.Order;
+import com.fastline.vendorservice.domain.model.Product;
+import com.fastline.vendorservice.domain.model.Vendor;
 import com.fastline.vendorservice.domain.repository.VendorRepository;
 import com.fastline.vendorservice.domain.service.VendorOrderService;
 import com.fastline.vendorservice.domain.service.VendorProductService;
 import com.fastline.vendorservice.domain.vo.VendorAddress;
 import com.fastline.vendorservice.domain.vo.VendorType;
+import com.fastline.vendorservice.infrastructure.external.dto.delivery.VendorHubIdMappingObj;
 import com.fastline.vendorservice.presentation.request.VendorCreateRequest;
 import com.fastline.vendorservice.presentation.request.VendorUpdateRequest;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
+import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -26,13 +33,10 @@ public class VendorService {
 	private final VendorProductService vendorProductService;
 	private final VendorOrderService vendorOrderService;
 
-	//    private final HubClient hubClient; -- 허브 서비스로 API 호출을 보낼 책임을 갖음
+	private final HubClient hubClient;
+	private final UserClient userClient;
 
-	/**
-	 * TODO: 허브 서비스로의 API 호출 성공, 실패시 흐름처리 작성 필요 TODO: 성공시, Vendor.create의 UUID.randomUUID() 부분에 허브ID.
-	 * 실패시 적절한 예외처리
-	 */
-	public Vendor insert(VendorCreateRequest createRequest) {
+	public Vendor insert(VendorCreateRequest createRequest, Long userId) {
 
 		VendorType vendorType = VendorType.fromString(createRequest.type());
 
@@ -43,10 +47,15 @@ public class VendorService {
 						createRequest.roadName(),
 						createRequest.zipCode());
 
-		//        hubClient.findHub(); -- 존재하는 허브인지 허브 서비스로 API 호출. 없거나 에러시 적절한 처리 필요
+		UUID passedHubId = validateHubId(createRequest.hubId());
 
 		Vendor vendor =
-				Vendor.create(createRequest.name(), vendorType, vendorAddress, UUID.randomUUID());
+				Vendor.create(createRequest.name(), vendorType, vendorAddress, passedHubId, userId);
+
+		UUID hubManagerId = userClient.getUserHubId(userId);
+		if (!isMasterUser(userId) && vendor.isHubManager(hubManagerId)) {
+			throw new CustomException(ErrorCode.VENDOR_FORBIDDEN);
+		}
 
 		return repository.insert(vendor);
 	}
@@ -59,27 +68,74 @@ public class VendorService {
 
 	@Transactional(readOnly = true)
 	public List<Product> findProductInVendor(UUID vendorId, Pageable pageable) {
+
 		return vendorProductService.findProductInVendor(vendorId, pageable);
 	}
 
 	@Transactional(readOnly = true)
-	public List<Order> findOrdersInVendor(UUID vendorId, Pageable pageable) {
+	public List<Order> findOrdersInVendor(UUID vendorId, Pageable pageable, Long userId) {
+
+		Vendor vendor = findByVendorId(vendorId);
+
+		UUID userHubId = userClient.getUserHubId(userId);
+		if (!isMasterUser(userId)
+				&& !vendor.isHubManager(userHubId)
+				&& !vendor.isVendorManager(userId)) {
+			throw new CustomException(ErrorCode.VENDOR_FORBIDDEN);
+		}
+
 		return vendorOrderService.findOrdersInVendor(vendorId, pageable);
 	}
 
-	/** TODO: hubId 업데이트 시도시, 유효한 Id인지 허브서비스로의 API요청 흐름 필요. TODO: 성공시 그대로 진행, 실패시 적절한 예외처리 */
-	public Vendor updateVendor(UUID vendorId, VendorUpdateRequest updateRequest) {
+	public Vendor updateVendor(UUID vendorId, VendorUpdateRequest updateRequest, Long userId) {
 
-		Vendor findVendor = repository.findByVendorId(vendorId);
-		if (updateRequest.hubId() != null && updateRequest.hubId() != findVendor.getHubId()) {
-			//            hubClient.findHub(); -- 존재하는 허브인지 허브 서비스로 API 호출. 없거나 에러시 적절한 처리 필요
+		Vendor vendor = repository.findByVendorId(vendorId);
+
+		UUID hubUserId = userClient.getUserHubId(userId);
+		if (!isMasterUser(userId)
+				&& !vendor.isHubManager(hubUserId)
+				&& !vendor.isVendorManager(userId)) {
+			throw new CustomException(ErrorCode.VENDOR_FORBIDDEN);
 		}
 
-		findVendor.update(updateRequest);
-		return repository.insert(findVendor);
+		if (updateRequest.hubId() != null && updateRequest.hubId() != vendor.getHubId()) {
+			validateHubId(updateRequest.hubId());
+		}
+
+		vendor.update(updateRequest);
+		return repository.insert(vendor);
 	}
 
-	public UUID deleteVendor(UUID vendorId) {
+	public UUID deleteVendor(UUID vendorId, Long userId) {
+
+		Vendor vendor = repository.findByVendorId(vendorId);
+
+		UUID userHubId = userClient.getUserHubId(userId);
+		if (!isMasterUser(userId) && !vendor.isHubManager(userHubId)) {
+			throw new CustomException(ErrorCode.VENDOR_FORBIDDEN);
+		}
+
 		return repository.deleteByVendorId(vendorId);
+	}
+
+	public List<UUID> findHubIdInVendor(UUID vendorSenderId, UUID vendorReceiverId) {
+		Map<UUID, UUID> mappingId =
+				repository.findHubIdInVendor(vendorSenderId, vendorReceiverId).stream()
+						.collect(
+								Collectors.toMap(VendorHubIdMappingObj::vendorID, VendorHubIdMappingObj::hubID));
+
+		return List.of(mappingId.get(vendorSenderId), mappingId.get(vendorReceiverId));
+	}
+
+	private UUID validateHubId(UUID hubId) {
+
+		if (hubClient.validateHubId(hubId)) {
+			throw new CustomException(ErrorCode.VENDOR_HUBID_INVALIDATION);
+		}
+		return hubId;
+	}
+
+	private boolean isMasterUser(Long userId) {
+		return userId == 1L;
 	}
 }
